refactored program into separate files, extracted game loop to new class
