+++ spec/commands_spec.rb
@@ -1,0 +1,12 @@
+require "commands"
+
+RSpec.describe Commands, "lets user quit or restart or save" do
+	let(:commands) { Class.new { extend Commands } }
+
+	describe "#leave" do 
+	  it "quits game when user enters \"y\"" do
+	  	answer = "y"
+	    expect(commands.leave).to eq 1
+	  end
+	end
+end
